+++ src/main/java/org/ua/fkrkm/proglatformdao/dao/ExerciseDaoI.java
@@ -5,5 +5,5 @@ import org.ua.fkrkm.proglatformdao.entity.Exercise;
 import java.util.List;
 
 public interface ExerciseDaoI extends ParentDaoI<Exercise> {
-    List<Exercise> findExercisesByTopicId(int topicId);
+    List<Exercise> findExercisesByTopicId(Long topicId);
 }
