+++ src/main/java/org/ua/fkrkm/proglatformdao/dao/ModuleDaoI.java
@@ -5,5 +5,5 @@ import org.ua.fkrkm.proglatformdao.entity.Module;
 import java.util.List;
 
 public interface ModuleDaoI extends ParentDaoI<Module> {
-    List<Module> getModulesByCourseId(int courseId);
+    List<Module> getModulesByCourseId(Long courseId);
 }
