+++ src/main/java/org/ua/fkrkm/proglatformdao/dao/ModuleStatDaoI.java
@@ -9,12 +9,12 @@ import java.util.List;
  * Інтерфейс для роботи з сутністю "Статистика по модулю"
  */
 public interface ModuleStatDaoI extends ParentDaoI<ModuleStat>{
-    List<ModuleStat> findModuleStatByUserId(Integer userId);
+    List<ModuleStat> findModuleStatByUserId(Long userId);
     /**
      * Отримати статистику по всім модулям по ID користувача
      *
      * @param userId ID користувача
      * @return List<ModuleStateView> статистика по модулям
      */
-    List<ModuleStateView> findModulesStatByUserId(Integer userId);
+    List<ModuleStateView> findModulesStatByUserId(Long userId);
 }
