+++ src/main/java/org/ua/fkrkm/proglatformdao/dao/impl/TestResultDaoImpl.java
@@ -26,7 +26,7 @@ public class TestResultDaoImpl extends ParentDaoImpl<TestResult> implements Test
     }
 
     @Override
-    public List<TestResult> getTestResultsByUserId(Integer userId) {
+    public List<TestResult> getTestResultsByUserId(Long userId) {
         String sql = "SELECT * FROM " + this.tableName + " WHERE user_id = :userId ORDER BY id DESC";
         SqlParameterSource sqlParameterSource = new MapSqlParameterSource()
                 .addValue("userId", userId);
