+++ src/main/java/org/ua/fkrkm/proglatformdao/mappers/CourseMapper.java
@@ -21,7 +21,7 @@ public class CourseMapper implements RowMapper<Course> {
     @Override
     public Course mapRow(ResultSet rs, int rowNum) throws SQLException {
         return Course.builder()
-                .id(rs.getInt("id"))
+                .id(rs.getLong("id"))
                 .name(rs.getString("name"))
                 .description(rs.getString("description"))
                 .created(rs.getTimestamp("created"))
